cmd/gg: add rebase tests
Found a bug where rebase would rebase on top of the source instead of
upstream.

Also discovered that the NoOp test is flaky, since it relies on the test
executing in under a second to keep the commit hashes the same.
Removed, since it's a bad test: it doesn't do anything fundamentally
different from the other tests.

Moved some of the test infrastructure around.