cmd/gg: fix rebase -src operation on unrelated revisions
Non-interactive `git rebase` does not support this, since it overloads
HEAD to both be the destination branch and the disambiguation path for
descendants specified by upstream. To work around this, gg kicks off an
"interactive" rebase that picks the right revisions.

To find descendants in this case, it examines all the local branches.

Fixing this uncovered that TestRebase was improperly using -src, so I
changed it to -base.

Fixes #27