internal/gittool: add StatusReader type
This was long overdue. Too many places replicated the status-read loop,
which is hard to get correct in the face of errors. This is especially
concerning in tests, which should be depending only on well-tested
functions. While the delta in this commit is large, the call sites are
much simpler to grok and less boilerplatey.

The new StatusReader wraps the same readStatusEntry function with a
bufio.Scanner-like API. The critical part is that calling Close() will
correctly send a kill signal to the subprocess and wait for it.
SignalReader suppresses EOFs and killed process errors so that errors
are more meaningful.