internal/gittool: surface placeholder for renamed file on buggy Git versions
Git 2.11 through Git 2.15 do not deal with renames well. git status will
produce malformed output and will not obey the diff.renames or
status.renames configuration settings. I work around this by
automatically detecting bad versions and changing the parse logic (with
heavy-handed comments and tests). When such a parsed entry is
encountered, gg status will exit with failure and a display a link to
the issue.

This is the most graceful way I could come up with surfacing this to the
user: most of the time they shouldn't hit it, and if they do, there's
messaging explaining the problem and the concrete action they can take
to address it (upgrade your Git version).

Fixes #60